-Cleaned up some debug and unused code.
Observation 1: ProspectUsers.vue could essentially be replaced by ProspectUsersTable.vue
It contains only contains that component and no props are being passed. I only did it that way in order to maintain the file structure you had, lmk if you want it changed.

Obervation 2: As for merging the ProspectUserStore into the LoggedInUserStore, I don't think the 2 stores are doing the same thing. The LoggedInUser seems to just keep track of the user who's logged in and other info related to the logged in user, whereas the ProspectUserStore sole purpose is to indicate whether there are prospect users present on the database or not. I believe merging the 2 will make the code less modular as they serve completely different purposes, but if you still want me to go ahead let me know and I'll do it.